refactor(sdk): split shielded mod into decrypt, notes_sync, and types
Move logical groups out of the monolithic shielded module:
- decrypt.rs: try_decrypt_note + MIN_ENCRYPTED_NOTE_LEN constant
- types.rs: ShieldedSyncConfig, DecryptedNote, ShieldedSyncResult
- notes_sync.rs: sync_shielded_notes + fetch_chunk
- nullifier_sync stays nested under shielded

mod.rs now just re-exports public items.
