+++ packages/rs-sdk/src/platform.rs
@@ -17,12 +17,13 @@ mod fetch_many;
 mod fetch_unproved;
 pub mod group_actions;
 pub mod identities_contract_keys_query;
-pub mod nullifier_sync;
 pub mod query;
 pub mod tokens;
 pub mod transition;
 pub mod trunk_branch_sync;
 pub mod types;
+#[cfg(feature = "shielded")]
+pub mod shielded;
 
 pub use dapi_grpc::platform::v0 as proto;
 pub use dash_context_provider::ContextProvider;
